Moving mock
Also adding load process as sink.
