--- tests/_files/build_datapackage_goal/data/elements/bus.csv
@@ -1,4 +1,0 @@
-name;type;balanced
-electricity;bus;True
-heat;bus;True
-ch4;bus;True
