--- tests/_files/build_datapackage_goal/data/periods/periods.csv
@@ -1,10 +1,0 @@
-timeindex;periods;timeincrement
-2016-01-01 00:00:00;0;1
-2016-01-01 01:00:00;0;1
-2016-01-01 02:00:00;0;1
-2030-01-01 00:00:00;1;1
-2030-01-01 01:00:00;1;1
-2030-01-01 02:00:00;1;1
-2050-01-01 00:00:00;2;1
-2050-01-01 01:00:00;2;1
-2050-01-01 02:00:00;2;1
