--- tests/_files/build_datapackage_goal/data/sequences/modex_tech_generator_gas_sequence.csv
@@ -1,1 +1,0 @@
-timeindex
