--- tests/_files/build_datapackage_goal/data/sequences/modex_tech_photovoltaic_utility_sequence.csv
@@ -1,10 +1,0 @@
-timeindex;photovoltaic_BB
-2016-01-01T00:00:00;0.0
-2016-01-01T01:00:00;0.0
-2016-01-01T02:00:00;0.16
-2030-01-01T00:00:00;0.0
-2030-01-01T01:00:00;0.0
-2030-01-01T02:00:00;0.3
-2050-01-01T00:00:00;0.0
-2050-01-01T01:00:00;0.0
-2050-01-01T02:00:00;0.5
