--- tests/_files/build_datapackage_goal/data/sequences/modex_tech_wind_turbine_onshore_sequence.csv
@@ -1,10 +1,0 @@
-timeindex;onshore_BB
-2016-01-01T00:00:00;0.0516
-2016-01-01T01:00:00;0.051
-2016-01-01T02:00:00;0.0444
-2030-01-01T00:00:00;0.0526
-2030-01-01T01:00:00;0.051
-2030-01-01T02:00:00;0.0444
-2050-01-01T00:00:00;0.0536
-2050-01-01T01:00:00;0.051
-2050-01-01T02:00:00;0.0444
